--- src/applications/user-testing/new-conditions/pages/followUp.js
@@ -1,189 +1,0 @@
-import {
-  currentOrPastMonthYearDateSchema,
-  currentOrPastMonthYearDateUI,
-  radioSchema,
-  radioUI,
-  selectSchema,
-  selectUI,
-  textareaUI,
-  titleUI,
-} from 'platform/forms-system/src/js/web-component-patterns';
-
-import { CONDITIONS_FIRST } from '../constants';
-import { conditionOptions } from '../content/conditionOptions';
-import { ServiceConnectedDisabilityDescription } from '../content/newConditions';
-import { causeOptions } from './conditionByConditionPages/cause';
-import { createItemName } from './conditionsFirstPages/utils';
-
-const getOtherConditions = (formData, currentIndex) => {
-  const otherNewConditions = formData?.conditionsFirst
-    ?.filter((_, index) => index !== currentIndex)
-    ?.map(item => createItemName(item));
-
-  return [...otherNewConditions];
-};
-
-const allCauses = ['NEW', 'SECONDARY', 'WORSENED', 'VA'];
-const causesWithoutSecondary = allCauses.filter(cause => cause !== 'SECONDARY');
-
-/** @type {PageSchema} */
-export default {
-  title: formData => createItemName(formData, true),
-  depends: formData => formData.demo === 'CONDITIONS_FIRST',
-  path: `new-conditions-${CONDITIONS_FIRST}-follow-up/:index`,
-  showPagePerItem: true,
-  arrayPath: 'conditionsFirst',
-  uiSchema: {
-    'ui:title': 'Conditions cause follow up',
-    conditionsFirst: {
-      items: {
-        ...titleUI(({ formData }) => createItemName(formData, true)),
-        date: currentOrPastMonthYearDateUI({
-          title: 'What’s the approximate date your condition started?',
-          hint: 'For example: January 2004 or 2004',
-        }),
-        cause: radioUI({
-          title: 'What caused your condition?',
-          labels: causeOptions,
-          options: {
-            updateSchema: (formData, _causeSchema, _causeUISchema, index) => ({
-              enum: getOtherConditions(formData, index).length
-                ? allCauses
-                : causesWithoutSecondary,
-            }),
-          },
-        }),
-        primaryDescription: textareaUI({
-          title:
-            'Briefly describe the injury or exposure that caused your condition. For example, I operated loud machinery while in the service, and this caused me to lose my hearing.',
-          required: (formData, index) =>
-            formData.conditionsFirst[index]?.cause === 'NEW',
-          charcount: true,
-          expandUnder: 'cause',
-          expandUnderCondition: 'NEW',
-        }),
-        'view:secondaryFollowUp': {
-          causedByCondition: selectUI({
-            title:
-              'Choose the service-connected disability that caused the new condition that you’re claiming here.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'SECONDARY' &&
-              getOtherConditions(formData, index).length > 0,
-            updateSchema: (formData, _schema, _uiSchema, index) => {
-              return selectSchema(getOtherConditions(formData, index));
-            },
-          }),
-          causedByConditionDescription: textareaUI({
-            title:
-              'Briefly describe how this disability caused your new condition.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'SECONDARY',
-            charcount: true,
-          }),
-          'ui:options': {
-            expandUnder: 'cause',
-            expandUnderCondition: 'SECONDARY',
-          },
-        },
-        'view:worsenedFollowUp': {
-          worsenedDescription: textareaUI({
-            title:
-              'Briefly describe the injury or exposure during your military service that caused your existing disability to get worse.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'WORSENED',
-            charcount: true,
-          }),
-          worsenedEffects: textareaUI({
-            title:
-              'Tell us how the disability affected you before your service, and how it affects you now after your service.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'WORSENED',
-            charcount: true,
-          }),
-          'ui:options': {
-            expandUnder: 'cause',
-            expandUnderCondition: 'WORSENED',
-          },
-        },
-        'view:vaFollowUp': {
-          vaMistreatmentDescription: textareaUI({
-            title:
-              'Briefly describe the injury or event while you were under VA care that caused your disability.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'VA',
-            charcount: true,
-          }),
-          vaMistreatmentLocation: textareaUI({
-            title: 'Tell us where this happened.',
-            required: (formData, index) =>
-              formData.conditionsFirst[index]?.cause === 'VA',
-            charcount: true,
-          }),
-          'ui:options': {
-            expandUnder: 'cause',
-            expandUnderCondition: 'VA',
-          },
-        },
-        'view:serviceConnectedDisabilityDescription': {
-          'ui:description': ServiceConnectedDisabilityDescription,
-        },
-      },
-    },
-  },
-  schema: {
-    type: 'object',
-    properties: {
-      conditionsFirst: {
-        type: 'array',
-        items: {
-          type: 'object',
-          required: ['cause'],
-          properties: {
-            date: currentOrPastMonthYearDateSchema,
-            cause: radioSchema(Object.keys(causeOptions)),
-            primaryDescription: {
-              type: 'string',
-              maxLength: 400,
-            },
-            'view:secondaryFollowUp': {
-              type: 'object',
-              properties: {
-                causedByCondition: selectSchema(conditionOptions),
-                causedByConditionDescription: {
-                  type: 'string',
-                  maxLength: 400,
-                },
-              },
-            },
-            'view:worsenedFollowUp': {
-              type: 'object',
-              properties: {
-                worsenedDescription: {
-                  type: 'string',
-                  maxLength: 50,
-                },
-                worsenedEffects: {
-                  type: 'string',
-                  maxLength: 350,
-                },
-              },
-            },
-            'view:vaFollowUp': {
-              type: 'object',
-              properties: {
-                vaMistreatmentDescription: {
-                  type: 'string',
-                  maxLength: 350,
-                },
-                vaMistreatmentLocation: {
-                  type: 'string',
-                  maxLength: 25,
-                },
-              },
-            },
-          },
-        },
-      },
-    },
-  },
-};
